Set `SlashableAttestationCheck` as an iface method.
And delete `CheckSlashableAttestation` from iface.
