+++ validator/client/attest_test.go
@@ -583,3 +583,153 @@ func TestServer_WaitToSlotOneThird_ReceiveBlockSlot(t *testing.T) {
 		t.Errorf("Wanted %d time for slot one third but got %d", uint64(time.Now().Unix()), currentTime)
 	}
 }
+
+func Test_slashableAttestationCheck(t *testing.T) {
+	for _, isSlashingProtectionMinimal := range [...]bool{false, true} {
+		t.Run(fmt.Sprintf("SlashingProtectionMinimal:%v", isSlashingProtectionMinimal), func(t *testing.T) {
+			validator, _, validatorKey, finish := setup(t, isSlashingProtectionMinimal)
+			defer finish()
+			var pubKey [fieldparams.BLSPubkeyLength]byte
+			copy(pubKey[:], validatorKey.PublicKey().Marshal())
+			att := &ethpb.IndexedAttestation{
+				AttestingIndices: []uint64{1, 2},
+				Data: &ethpb.AttestationData{
+					Slot:            5,
+					CommitteeIndex:  2,
+					BeaconBlockRoot: bytesutil.PadTo([]byte("great block"), 32),
+					Source: &ethpb.Checkpoint{
+						Epoch: 4,
+						Root:  bytesutil.PadTo([]byte("good source"), 32),
+					},
+					Target: &ethpb.Checkpoint{
+						Epoch: 10,
+						Root:  bytesutil.PadTo([]byte("good target"), 32),
+					},
+				},
+			}
+
+			err := validator.db.SlashableAttestationCheck(context.Background(), att, pubKey, [32]byte{1}, false, nil)
+			require.NoError(t, err, "Expected allowed attestation not to throw error")
+		})
+	}
+}
+
+func Test_slashableAttestationCheck_UpdatesLowestSignedEpochs(t *testing.T) {
+	for _, isSlashingProtectionMinimal := range [...]bool{false, true} {
+		t.Run(fmt.Sprintf("SlashingProtectionMinimal:%v", isSlashingProtectionMinimal), func(t *testing.T) {
+			validator, m, validatorKey, finish := setup(t, isSlashingProtectionMinimal)
+			defer finish()
+			ctx := context.Background()
+			var pubKey [fieldparams.BLSPubkeyLength]byte
+			copy(pubKey[:], validatorKey.PublicKey().Marshal())
+			att := &ethpb.IndexedAttestation{
+				AttestingIndices: []uint64{1, 2},
+				Data: &ethpb.AttestationData{
+					Slot:            5,
+					CommitteeIndex:  2,
+					BeaconBlockRoot: bytesutil.PadTo([]byte("great block"), 32),
+					Source: &ethpb.Checkpoint{
+						Epoch: 4,
+						Root:  bytesutil.PadTo([]byte("good source"), 32),
+					},
+					Target: &ethpb.Checkpoint{
+						Epoch: 10,
+						Root:  bytesutil.PadTo([]byte("good target"), 32),
+					},
+				},
+			}
+
+			m.validatorClient.EXPECT().DomainData(
+				gomock.Any(), // ctx
+				&ethpb.DomainRequest{Epoch: 10, Domain: []byte{1, 0, 0, 0}},
+			).Return(&ethpb.DomainResponse{SignatureDomain: make([]byte, 32)}, nil /*err*/)
+			_, sr, err := validator.getDomainAndSigningRoot(ctx, att.Data)
+			require.NoError(t, err)
+
+			err = validator.db.SlashableAttestationCheck(context.Background(), att, pubKey, sr, false, nil)
+			require.NoError(t, err)
+			differentSigningRoot := [32]byte{2}
+
+			err = validator.db.SlashableAttestationCheck(context.Background(), att, pubKey, differentSigningRoot, false, nil)
+			require.ErrorContains(t, "could not sign attestation", err)
+
+			e, exists, err := validator.db.LowestSignedSourceEpoch(context.Background(), pubKey)
+			require.NoError(t, err)
+			require.Equal(t, true, exists)
+			require.Equal(t, primitives.Epoch(4), e)
+			e, exists, err = validator.db.LowestSignedTargetEpoch(context.Background(), pubKey)
+			require.NoError(t, err)
+			require.Equal(t, true, exists)
+			require.Equal(t, primitives.Epoch(10), e)
+		})
+	}
+}
+
+func Test_slashableAttestationCheck_OK(t *testing.T) {
+	for _, isSlashingProtectionMinimal := range [...]bool{false, true} {
+		t.Run(fmt.Sprintf("SlashingProtectionMinimal:%v", isSlashingProtectionMinimal), func(t *testing.T) {
+			ctx := context.Background()
+			validator, _, _, finish := setup(t, isSlashingProtectionMinimal)
+			defer finish()
+			att := &ethpb.IndexedAttestation{
+				AttestingIndices: []uint64{1, 2},
+				Data: &ethpb.AttestationData{
+					Slot:            5,
+					CommitteeIndex:  2,
+					BeaconBlockRoot: []byte("great block"),
+					Source: &ethpb.Checkpoint{
+						Epoch: 4,
+						Root:  []byte("good source"),
+					},
+					Target: &ethpb.Checkpoint{
+						Epoch: 10,
+						Root:  []byte("good target"),
+					},
+				},
+			}
+			sr := [32]byte{1}
+			fakePubkey := bytesutil.ToBytes48([]byte("test"))
+
+			err := validator.db.SlashableAttestationCheck(ctx, att, fakePubkey, sr, false, nil)
+			require.NoError(t, err, "Expected allowed attestation not to throw error")
+		})
+	}
+}
+
+func Test_slashableAttestationCheck_GenesisEpoch(t *testing.T) {
+	for _, isSlashingProtectionMinimal := range [...]bool{false, true} {
+		t.Run(fmt.Sprintf("SlashingProtectionMinimal:%v", isSlashingProtectionMinimal), func(t *testing.T) {
+			ctx := context.Background()
+			validator, _, _, finish := setup(t, isSlashingProtectionMinimal)
+			defer finish()
+			att := &ethpb.IndexedAttestation{
+				AttestingIndices: []uint64{1, 2},
+				Data: &ethpb.AttestationData{
+					Slot:            5,
+					CommitteeIndex:  2,
+					BeaconBlockRoot: bytesutil.PadTo([]byte("great block root"), 32),
+					Source: &ethpb.Checkpoint{
+						Epoch: 0,
+						Root:  bytesutil.PadTo([]byte("great root"), 32),
+					},
+					Target: &ethpb.Checkpoint{
+						Epoch: 0,
+						Root:  bytesutil.PadTo([]byte("great root"), 32),
+					},
+				},
+			}
+
+			fakePubkey := bytesutil.ToBytes48([]byte("test"))
+			err := validator.db.SlashableAttestationCheck(ctx, att, fakePubkey, [32]byte{}, false, nil)
+			require.NoError(t, err, "Expected allowed attestation not to throw error")
+			e, exists, err := validator.db.LowestSignedSourceEpoch(context.Background(), fakePubkey)
+			require.NoError(t, err)
+			require.Equal(t, true, exists)
+			require.Equal(t, primitives.Epoch(0), e)
+			e, exists, err = validator.db.LowestSignedTargetEpoch(context.Background(), fakePubkey)
+			require.NoError(t, err)
+			require.Equal(t, true, exists)
+			require.Equal(t, primitives.Epoch(0), e)
+		})
+	}
+}
