+++ validator/db/filesystem/BUILD.bazel
@@ -23,8 +23,10 @@ go_library(
         "//validator/db/kv:go_default_library",
         "@com_github_ethereum_go_ethereum//common/hexutil:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
+        "@com_github_prometheus_client_golang//prometheus:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@in_gopkg_yaml_v3//:go_default_library",
+        "@io_opencensus_go//trace:go_default_library",
     ],
 )
