+++ validator/db/iface/BUILD.bazel
@@ -18,5 +18,6 @@ go_library(
         "//proto/prysm/v1alpha1:go_default_library",
         "//validator/db/filesystem:go_default_library",
         "//validator/db/kv:go_default_library",
+        "@com_github_prometheus_client_golang//prometheus:go_default_library",
     ],
 )
